Work out more structure
